refactor(observeOn): move observeOn support classes to own file for sharing
